ingestion/requirements.txt
# Document Processing
pdfplumber==0.11.0
PyMuPDF==1.23.26
python-docx==1.1.0
unstructured==0.11.8
pytesseract==0.3.10
Pillow==10.2.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Storage
boto3==1.34.31
minio==7.2.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-magic==0.4.27
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
loguru==0.7.2

# Async
aiofiles==23.2.1
httpx==0.26.0
